Add files via upload
This model predicts the recommended movies for a user. It is basically using Natural Language Processing that uses difflib, TfidfVectorizer and cosine_similarity to predict the most similar movies according to user's input. 
difflib is used in case of spelling or grammatical mistakes by the user in giving the movie name this will find the nearest correct movie name.
TfidfVectorizer is the base building block of many NLP pipelines. It is a simple technique to vectorize text documents i.e. transform sentences into arrays of numbers and use them in subsequent tasks. 
cosine_similarity is used to measure the similarity between two vectors of an inner product space. It is measured by the cosine of the angle between two vectors and determines whether two vectors are pointing in roughly the same direction. It is used here to measure document similarity in text analysis.
Used Pandas and Numpy for Data collection and Preprocessing.
The model predicts 30 similar movies to what the user has given input.